Crate serde_human_bytes

Serialize byte arrays and vectors as bytes or as human-readable strings, depending on the format.

For example, serialize as hex if serializing to JSON, or as an efficient byte sequence with CBOR.

Modules

base64_vec

Serialize a byte vector as base64 if human-readable, or as bytes if not.

hex_array

Serialize a byte array as a list of bytes if human-readable, or as hex if not.

Structs

Base64Vec

A byte vector that serializes as base64 in human-readable formats.

HexArray

A byte array that serializes as hex in human-readable formats.